Fix a few things that linting broke

+++ api/brep/HalfEdge.h
@@ -5,33 +5,35 @@
 
 #pragma once
 
+#include <cstdint>
+
 namespace v3d::brep {
 
     class HalfEdge {
      public:
         HalfEdge();
-        explicit HalfEdge(unsigned int vert);
+        explicit HalfEdge(uint64_t vert);
         explicit HalfEdge(const HalfEdge & e);
         ~HalfEdge();
 
         bool operator == (const HalfEdge & e);
         HalfEdge & operator = (const HalfEdge & e);
 
-        unsigned int vertex(void) const;
-        unsigned int face(void) const;
-        unsigned int pair(void) const;
-        unsigned int next(void) const;
+        uint64_t vertex(void) const;
+        uint64_t face(void) const;
+        uint64_t pair(void) const;
+        uint64_t next(void) const;
 
-        void vertex(unsigned int vert);
-        void face(unsigned int f);
-        void pair(unsigned int e);
-        void next(unsigned int e);
+        void vertex(uint64_t vert);
+        void face(uint64_t f);
+        void pair(uint64_t e);
+        void next(uint64_t e);
 
      private:
-        unsigned int vertex_;  // vertex at end of half edge
-        unsigned int face_;  // face to left of edge
-        unsigned int pair_;  // symetric half edge
-        unsigned int next_;  // next CCW half edge
+        uint64_t vertex_;  // vertex at end of half edge
+        uint64_t face_;  // face to left of edge
+        uint64_t pair_;  // symetric half edge
+        uint64_t next_;  // next CCW half edge
     };
 
 };  // namespace v3d::brep
